requirements.txt
python-dotenv==1.0.1
requests==2.31.0
aiohttp==3.9.3
discord.py==2.3.2


